updated all file anmes and routes and created files for titles table and create function.
