+++ controllers/departments/getall.js
@@ -2,17 +2,16 @@ const mysql = require('mysql')
 const pool = require("../../sql/connection")
 const { handleSQLError } = require('../../sql/error')
 
-
 const getDepartments = (req, res) => {
-    let sql = ` SELECT *
-                FROM departments`
-    sql = mysql.format(sql, [])
-    pool.query(sql, (err, rows) => {
-        if (err) return handleSQLError(res, err)
-        return res.json(rows);
-    })
+  let sql = ` SELECT * FROM departments`
+  sql = mysql.format(sql, [])
+
+  pool.query(sql, (err, rows) => {
+      if (err) return handleSQLError(res, err)
+      return res.json(rows);
+  })
 }
 
 module.exports = {
-  getDepartments 
+  getDepartments
 }
