+++ controllers/employees/create.js
@@ -1,0 +1,3 @@
+// const mysql = require('mysql')
+// const pool = require("../../sql/connection")
+// const { handleSQLError } = require('../../sql/error')
