+++ index.js
@@ -2,7 +2,7 @@ const express = require("express")
 const bodyParser = require("body-parser")
 const employeesRouter = require("./routes/employees")
 const departmentsRouter = require("./routes/departments")
-const titlesRouter = require("./routes/titles")
+// const titlesRouter = require("./routes/titles")
 const salariesRouter = require("./routes/salaries")
 
 require('dotenv').config()
@@ -12,7 +12,7 @@ const port = process.env.PORT || 4001
 app.use(bodyParser.json())
 app.use('/employees', employeesRouter)
 app.use('/departments', departmentsRouter)
-app.use('/titles', titlesRouter)
+// app.use('/titles', titlesRouter)
 app.use('/salaries', salariesRouter)
 
 app.get('/', (req, res) => {
@@ -21,5 +21,5 @@ app.get('/', (req, res) => {
 
 app.listen(port, () => {
  console.log(`Web server is listening on port ${port}!`)
-}) 
+})
