+++ testing/titleControllerTest.js
@@ -1,0 +1,33 @@
+// let titlesGetController = require("../controllers/titles/getAll");
+// let titlesGetIdController = require("../controllers/titles/getById");
+// let titlesDeleteController = require("../controllers/titles/delete");
+// let titlesUpdateController = require("../controllers/titles/update");
+// let titlesCreateController = require("../controllers/titles/create");
+
+// let mockResponse = {
+//   json: (body) => {
+//     console.log(body);
+//     return mockResponse;
+//   },
+//   send: (body) => {
+//     console.log(body);
+//     return mockResponse;
+//   },
+//   status: (code) => {
+//     console.log(code);
+//     return mockResponse;
+//   },
+// };
+
+// let mockRequest = {
+//   // TODO: Need to update below code
+//   // params: { dept_no: "d009"},
+//   // body: { dept_no: "d011", dept_name: "fish killers" },
+// };
+
+// //UNCOMMENT WHATEVER YOU ARE TRYING TO TEST
+// titlesGetController.getTitles(mockRequest,mockResponse);
+// // titlesGetIdController.getTitlesByNo(mockRequest,mockResponse);
+// // titlesCreateController.createTitle(mockRequest,mockResponse);
+// // titlesUpdateController.updateTitle(mockRequest,mockResponse);
+// // titlesDeleteController.deleteTitle(mockRequest,mockResponse);
